Core Script
This script will not check for existing K2 images; any existing images will be overwritten. It's best to run this immediately after importing Joomla! content into K2.  Simply place the file in your Joomla root, in the same folder as configuration.php. The script uses the same DB parameters as Joomla. This is a basic framework script; if you have a lot of articles, or the script times out, you might need to adjust the script to track articles that have been processed, or only copy and process images if they do not exist. TWEAK THIS FILE AS NECESSARY to suit your needs.